DEV-35909 fix type of year field in invoice number counter table (#556)

+++ src/main/java/de/kfzteile24/salesOrderHub/domain/InvoiceNumberCounter.java
@@ -19,7 +19,7 @@ import javax.persistence.Table;
 public class InvoiceNumberCounter {
 
     @Id
-    @Column(name = "year", columnDefinition = "smallint", updatable = false)
+    @Column(name = "year", columnDefinition = "int", updatable = false)
     private Integer year;
 
     private Long counter;
